@@ -24,10 +24,10 @@
     "local-storage": "^2.0.0",
     "near-api-js": "^0.45.1",
     "prettier": "^2.7.1",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
     "react-bootstrap-typeahead": "^6.0.0",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^3.1.4",
     "react-files": "^3.0.0-alpha.3",
     "react-infinite-scroller": "^1.2.6",
